+++ server/src/services/types.d.ts
@@ -1,0 +1,13 @@
+export enum Platform {
+  YOUTUBE,
+  SOUNDCLOUD,
+}
+
+export interface Resource {
+  title: string;
+  length: number;
+  author: string;
+  thumbnail: string;
+  url: string;
+  platform: Platform;
+}
